Add WebSocket provider and events Eth & Nervos tests.
