+++ hardhat.config.eth.js
@@ -1,0 +1,9 @@
+require('@nomiclabs/hardhat-web3');
+
+module.exports = {
+    solidity: '0.8.4',
+    paths: {
+        artifacts: "./src/artifacts",
+        tests: "./test-eth"
+    }
+};
