--- src/config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    HTTP_RPC_URL: 'https://godwoken-testnet-web3-rpc.ckbapp.dev',
-    USER_ONE_PRIVATE_KEY: '0xd9066ff9f753a1898709b568119055660a77d9aae4d7a4ad677b8fb3d2a571e5'
-};
